create tests for utils module
